SCHED-340: Modifications suggested by PR review.
